requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Supabase
supabase==2.10.0

# HTTP client
httpx==0.27.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Blockchain verification (settlement report timestamps)
opentimestamps==0.3.0

# Email (Phase 2)
sendgrid==6.11.0

# PDF Generation (Phase 2)
weasyprint==63.1
pillow==11.0.0
qrcode[pil]==8.0

# File Storage (Phase 2)
boto3==1.35.99
botocore==1.35.99

# Payment Processing (Phase 2)
stripe==11.3.0

# Monitoring (Phase 2)
sentry-sdk[fastapi]==2.19.2

# Utilities
python-dotenv==1.0.1
pytz==2024.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
